Moved onSuccess call after render call
In order to get the dimensions (e.g. clientHeight, clientWidth) of the embed it needs to be rendered to the screen. This is why the onSuccess call is better made after the internal state update that causes the rerender.
Alternatively there could be an additional callback such as `onRender`. But since I didn’t see any particular advantage of placing the onSuccess call before the render, it is easier to just move it to the end. Hope this makes sense :)